issue: EmailTest Draft
This addresses a very minor issue where the draft save on EmailTest throws
an error. This is due to changes made in 5694 where `message` was removed
from `$field_list` (security reasons). This updates the name of the textarea
from `message` to `body`. This allows the `_findDraftBody()` method to find
the body successfully and continue on. This also updates every reference of
`message` to `body` in this instance.