issue: Edit Entry Dropped Attachments
This addresses an issue where editing your own Thread Entry and clicking
Save two separate times drops the attachment(s). This is due to the order of
operations in the backend where if editing your own Thread Entry we delete
the previous entry and set `$old` to the original Thread Entry (parent of
the previous entry). This causes the attachment lookup to search for the
wrong `object_id` which obviously finds no attachments. This moves the
attachment lookup before we delete the previous entry so we can search with
the appropriate ID.