Issue: Missing null check in Staff->updatePerms()
If a agent has no permissions and you will give him/her the first permission, the class Staff->updatePerms() wrote a `array_key_exists() expects parameter 2 to be array, null given` warning.

Problem is, that $permissions->perms is null. With a RolePermission->exists() function is the code easier to read an don't show the warning.