Issue: Queue Columns Custom Fields
This commit fixes an issue where a DepartmentField or PriorityField added to a Custom Form would not show up in a
 Queue Column.

PriorityField:
In to_php, if the $value passed in is actually the priority_id or the $id passed in is actually the priority_id, we should use the id to get the Priority value that should be displayed, however, if the $value passed in is already the Priority value as a string, we should return that value.

DepartmentField:
A display function needed to be added for this field. If the $dept passed in is the Department name, we should return that, otherwise if the $dept isn't a Dept object, we should use the parent display function as it was doing before. Finally, if the $dept is a Dept object, we should get the Department name from that.