Issue: Filter Email Variables
This commit fixes an issue where some template variables could not be used when using a Filter to send an email. The filter should run after the appropriate variables have been created in the ticket, and we need to set the recipient as the ticket owner within the filter action as well.