issue: Missing Thread On Referral Check
This addresses an issue where attempting to Assign a Child Ticket throws a
fatal error of `Call to a member function getReferral() on null`. This is
due to the `thread` relation not being available for Child Tickets; instead
`child_thread` is available. This updates the code to use `getThread()`
which first checks if `thread` relation exists, if not it returns the
`child_thread` relation.