issue: getDBVersion() SQL Errors
This addresses an issue where when `OsticketConfig::getDBVersion()` is ran,
a `column not found` SQL Error is generated. This is due to the config table
structure being completely different from 1.7 onwards. This adds a check to
see if the `MAJOR_VERSION` constant is defined, if so, it will skip the
`OsticketConfig::getDBVersion()` call in `include/class.ostsession.php`
(thus skipping the query execution). If the constant is not defined it will
continue on and call `OsticketConfig::getDBVersion()`.