session: Verify UserAgent String
This adds an extra layer of security to Sessions where we will verify the
UserAgent string. We will check to see if the UserAgent string saved in the
database matches the current UserAgent string from the request. If the two
do not match we will destroy the session and force the affected User/Agent
to login again.