xss: Client-Side i18n
This mitigates a vulnerability reported by @tieupham267 where XSS is
possible via the `REQUEST_URI` parameter on the Client Portal. This is due
to the parameter not being converted to HTML entities before attaching to
`<link>` tags. This adds `htmlspecialchars()` around the `REQUEST_URI` to
ensure the content is properly converted.