security: open.php Refresh
This addresses an issue reported by Vladislav Alyohin where Users can create
a Ticket in the Client Portal and click Refresh over and over again on
success. The result is a duplicate Ticket being created (with a different
Ticket Number) for each refresh. This adds an else statement so if there is
no client or if they are no longer valid we will rotate the CSRF token. This
will create the initial Ticket and on refresh will force the User back to
the Landing Page.