issue: Setup Admin Password Heltip Verbiage
This addresses 5930 where the Helptip for the Admin Password field on
Install is incorrect. The Helptip states the password has to be a minimum of
5 characters where in reality it requires a minimum of 6. This updates the
Helptip as well as the inline string from `Must be five (5) characters or
more.` to `Must be six (6) characters or more.`. In addition, this updates
the phrase `(Five characters min)` to `(Six characters min)` in
`include/class.validator.php` as well as updates the check for less than 5
characters.