Feature: Filter Thread Event Improvement
This commit improves upon the feature by adding a 'getEventDescription' function to the 'TriggerAction' class so that we don't have to check to see if the method exists for the type of filter action.

It also ensures that an object is found that can be used within the Thread Event, and if we don't find one, we will just generalize the wording.

Ex:
Thread Event where name is found:
Set Department Filter set Department to Parent

Thread Event where Dept name is not found:
Set Department Filter set Department