search: Child Thread Relation
This addresses issue 5959 where `TicketThread / Last Message` and
`TicketThread / Last Response` appear twice in the Advanced Search available
Fields. This is due to the joins/relationships automatically being added to
searchable fields. In this case, searching TicketThread will search the
Parent and Child Threads so there is no need for two separate options. This
adds a new property called `searchable` (set to false) to the `child_thread`
relation in the Ticket class. In addition, this adds a check when generating
the searchable fields that skips the relation if the `searchable` property
is `false`.