issue: def_assn_role
This addresses an issue reported on the Forum where disabling `Fall back to
primary role on assignments` doesn’t actually save. This is due to the third
parameter passed being `false`. The third parameter is `$commit` that
determines wether or not to save it after changing. Since this was `false`
it was never committed/saved in the database. This updates the parameter to
`true` so that it gets committed/saved.