issue: Auto-Assign Comments Var
This addresses 5978 where when a Filter auto-assigns an Agent, the New
Assignment Alert they receive shows the `%{comments}` variable. This is due
to `$comments` being `false` therefore it skips the variable. This adds a
check for `$comments` and if empty/false uses `''` so the variable is
replaced/removed from the final alert.