issue: CDATA Rebuild Issue
This addresses issues 6043 and 6053 where installing osTicket and doing a
User lookup will return a fatal error of `Unknown column 'A5.phone' in
'where clause'`. This is an indirect result of the changes we did to rebuild
CDATA tables post-installation. The issue is that we are using
`getInstance()` which returns a cached object. Changing this to always
return a new instance via `getNewInstance()` subsequently had issues with
the TicketForm not adding the Priority field. We determined this is due to
object caching in the ORM so we added `ModelInstanceManager::flushCache()`
before we rebuild the CDATA tables to clear the object cache and prevent the
reuse of old objects.