issue: Assignee Item Property
This addresses an issue where you cannot save a value for an Assignee field
that’s in the Item Properties of a List Item. This is due to the method used
to prepare the data for the backend as well as the format of the data when
returned from `to_database()`. This updates `to_php()` to `to_database()` to
get the appropriate data for the backend. This also updates `to_database()`
to return the JSON encoded data itself instead of in an array. Lastly, this
updates the `getChanges()` method to remove the need to account for an
array.