issue: Status List getSortModes()
This addresses issue 6063 where TicketStatusList does not have a
`getSortModes()` method and fatal fails when loading the status list. This
was done by `__call()` via `CustomListHandler` previously but we changed
`getSortModes()` to a static method to match PHP 8 standards. We didn't
account for the fact that `__call()` cannot call a static method as it uses
`$this` context. This renames the `getSortModes()` method for `DynamicList`
to `sortModes()`, creates a new method called `getSortModes()` which returns
the self `sortModes()`, updates `dynamic-list.inc.php` to call the method
non-statically and for `DynamicList` calls `sortModes()` statically, and
removes `static` keyword from `getSortModes()` on `CustomList` interface.