issue: PEAR Mail parseAddressList()
This addresses an issue reported on the Forum where when
`Mail::parseRecipients()` is called it throws a fatal error of `Uncaught
ValueError: mb_convert_encoding(): Argument #2 ($to_encoding) must be a
valid encoding, "localhost" given`. This is due to an erroneous change when
adding PHP 8.0 support. We changed the called class from `Mail_RFC822` to
`Mail_Parse` but left the same arguments. This is obviously an issue as both
classes' methods take different arguments. This changes the called class
back to `Mail_RFC822` to stick to original functionality and updates it for
PHP 8.0 standards.