issue: Filter Events
This addresses issue 6059 where Ticket Filter processing stops accurately
but the Thread Events are blindly added for any Filters that match the
criteria. This is due to not checking the current Filter in the iteration
for `stopOnMatch()`. This adds a check for `stopOnMatch()` and if `true` we
will break the loop.