issue: Last Message Data Source
This addresses an issue reported on the Forum where the Last Message and
Last Response Data Sources are not showing the correct times. This is due to
an erroneous change made to support PHP 8.0. We added `isset()` to parts
where we are selecting array keys but didn’t add a check for the value. This
adds checks for the values to ensure we are continuing at the appropriate
times.