issue: Nullable date() Timestamp
This addresses an issue where starting with PHP 8.0 the `date()` method
changed the `timestamp` parameter to be nullable. If the `timestamp` is
`NULL` then it will return the output from the `time()` method. So instead
of returning a default date of `1970-01-01T00:00:00+00:00` like it used to
it returns the current date. This is obviously a problem when it comes to
datetime based Queue Columns like Last Response with no value. Since there
is no value it gets translated to the current time and displays `just now`
in the columns instead of a blank column. This updates the `date()` call to
use `0` if the `timestamp` is `NULL` so that it returns the default datetime
of `1970-01-01T00:00:00+00:00` essentially restoring previous behavior.