xss: Agent Directory
This mitigates a vulnerability discovered by the AppSec Research Team at
Checkmarx where it's possible to perform Reflected XSS in the Agent
Directory. This was due to the `order` and `sort` URL params not being
properly escaped. This adds `Format::htmlchars()` to both params to ensure
they are properly escaped.