xss: System Logs
This mitigates a vulnerability discovered by the AppSec Research Team at
Checkmarx where it's possible to perform Stored XSS via invalid CSRF tokens
in the System Logs. This is due to not properly sanitizing the CSRF content
before saving to the database as well as not properly escaping the content
before displaying in the UI. This adds `Format::sanitize()` to the content
before saving to the database as well as `Format::htmlchars()` before
displaying.