security: Unvalidated Password Change
This mitigates a vulnerability discovered by the AppSec Research Team at
Checkmarx where it's possible to update your Agent password without passing
a valid current password. This adds a check to compare the provided current
password with the actual current password to ensure they match. If they do
not match an error will be returned stating `Current password is
incorrect.`.