issue: Redactor Freezing
This addresses an issue where removing all text in Redactor in the Post
Reply box freezes the entire page (using 90% or more CPU). As a quick
workaround user jarnott from the Forum found that adding a `br` tag
between the `draft_id` input and the Post Reply `textarea` fixes the
issue. This adds the `br` tag to temporarily address the issue until
Redactor addresses the issue permanently.