installer: Help Topic Disabled Fields
This addresses an issue where installing `1.16.x` has the Ticket Details
Form Fields disabled for every Help Topic. This is due to checking if the
Field IDs are in an empty array which obviously will return false. This
updates the declaration of `$fields` from `array()` to `null` if the array
index doesn’t exist. This also updates the `if` statement for the Field ID
checking to ensure the `$fields` variable is an array. This ensures that the
check is skipped if `$fields` is `null` avoiding erroneous disabled fields.