issue: Check User Status
This is a rewrite of 6163 where we aren’t checking User/Agent status on
login and on session validation. This adds checks to the
UserAuthenticationBackend to ensure the User is active (ie. confirmed and
not administratively locked). This also adds checks to the
StaffAuthentciationBackend to ensure the Agent is active (ie. not locked).
If a User/Agent is not active they will be redirected to the login screen.