update: More Lookup Improvements
This adds a few more lookup improvements. This adds a new Validator called
`is_alnumeric()` that will check if the provided string is in fact
alphanumeric. This also updates an if statement in
`TicketsAjaxAPI::lookup()` to add a check for `is_alnumeric()` to make sure
we are only looking up Ticket numbers with `lookupByNumber()`.