plugins: Make Plugins Upgradeable Again!
This adds ability and interface to auto-trigger plugins upgrade or downgrade
when there's a version mismatch on-load.

The base plugin class simply updates the version (for now) - the downstream
Plugins should implement the actual upgrade / downgrade process as needed,
especially if it creates database tables or needs to change stored config info
to match the new version format.