plugins: Multi-Instance Blacklist
Some Plugins DO NOT or SHOULDN'T support multiple instances due design issues
or simply because of the fact that it doesn't make sense to do so.

The blacklist will be removed down the road once plugins are forced to declare
compatiple core osTicket version and if it supports multiple instances.