mailfetcher: Catch & Log Post Create Errors
This commit adds a try/catch to catch possible exceptions that might happen
post ticket create. This is necessary so we can isolate problematic emails from
segfaulting the entire process (fatal error) or cause it to skip expunging the
mailbox.

The commit also adds debug logging for such errors / exceptions.